--- cerebro-temp/app.py
@@ -1,206 +1,0 @@
-import os
-
-from flask import Flask, render_template
-from flask_sqlalchemy import SQLAlchemy as sql
-import hashlib
-from helpers import *
-import json
-
-app = Flask(__name__)
-
-app.secret_key = os.environ["SECRET_KEY"]
-
-try:
-    app.config["SQLALCHEMY_DATABASE_URI"] = os.environ["DATABASE_URI"]
-
-except KeyError:
-    raise ValueError("É necessário definir variável de ambiente DATABASE_URI com a chave de acesso do banco de dados.")
-
-db = sql(app)
-start_db(db)
-
-@app.route("/access", methods=["GET", "POST"])
-def access():
-    if request.method == "POST":
-        name = request.form["name"]
-        hash_result = hashlib.sha256(request.form["key"].encode()).hexdigest()
-        result = db.engine.execute("SELECT name, permissions FROM users INNER JOIN roles ON users.role_id=roles.id WHERE name = %s AND password = %s", name, hash_result).first()
-        if result:
-            session["user_id"] = result[0]
-            session["roles"] = result[1]
-            return redirect("/")
-        else:
-            flash("Username or password incorrect")
-    return render_template("access.html")
-
-@app.route("/")
-@login_required
-@admin_role
-def index():
-    contracts = db.engine.execute("SELECT COUNT(*) FROM services").first()[0]
-    no_payments_contracts = db.engine.execute("SELECT COUNT(*) FROM services WHERE first_payment is null").first()[0]
-    no_accepted_contracts = db.engine.execute("SELECT COUNT(*) FROM services WHERE acceptance_date is null").first()[0]
-    clients = db.engine.execute("SELECT COUNT(*) FROM clients").first()[0]
-    return render_template("index.html", contracts=contracts, no_payments_contracts=no_payments_contracts, no_accepted_contracts=no_accepted_contracts, clients=clients)
-
-@app.route("/logout")
-@login_required
-def logout():
-    session["user_id"] = None
-    session["roles"] = None
-    return redirect("/")
-
-@app.route("/contracts_manager")
-@login_required
-@admin_role
-def contracts_manager():
-    raw_contracts = db.engine.execute("SELECT services.*, count(payments.service_id) FROM services LEFT JOIN payments ON services.id=payments.service_id group by services.id, payments.id")
-    contracts = []
-    for contract in raw_contracts:
-        short_description = contract[7]["short_description"]
-        service_list = contract[7]["service_list"]
-        contracts.append(list(contract[:7]) + [short_description, service_list] + list(contract[8:]))
-    return render_template("contracts_manager.html", contracts=contracts)
-
-@app.route("/clients_manager")
-@login_required
-@admin_role
-def clients_manager():
-    clients = db.engine.execute("SELECT * FROM clients ORDER BY id")
-    return render_template("clients_manager.html", clients=list(clients))
-
-@app.route("/access_manager")
-@login_required
-@admin_role
-def access_manager():
-    raw_accounts = db.engine.execute("SELECT name, role_id, permissions FROM users INNER JOIN roles on role_id=id")
-    accounts = [[name, role, " ".join(permissions)] for name, role, permissions in raw_accounts]
-    roles = db.engine.execute("SELECT * FROM roles")
-    return render_template("access_manager.html", accounts=list(accounts), roles=list(roles))
-
-@app.route("/ata_manager")
-@login_required
-@admin_role
-def ata_manager():
-    atas = db.engine.execute("SELECT * FROM ej ORDER BY ata_date")
-    return render_template("ata_manager.html", atas=list(atas))
-
-@app.route("/access_creation", methods=["POST"])
-@login_required
-@admin_role
-def access_creation():
-    name = request.form['username']
-    hash = hashlib.sha256(request.form['password'].encode()).hexdigest()
-    role_id = request.form['role']
-    db.engine.execute("INSERT INTO users(name, password, role_id) VALUES(%(name)s, %(hash)s, %(role)s)", name=name, hash=hash, role=role_id)
-    return redirect(url_for("access_manager"))
-
-@app.route("/role_creation", methods=["POST"])
-@login_required
-@admin_role
-def role_creation():
-    permissions = normalize_array(request.form['permissions'])
-    db.engine.execute("INSERT INTO roles(permissions) VALUES(%(p)s)", p=permissions)
-    return redirect(url_for("access_manager"))
-
-@app.route("/edit_accounts", methods=["POST"])
-@login_required
-@admin_role
-def edit_accounts():
-    complete_request = json.loads(request.get_data().decode("utf-8"))
-    for req in complete_request:
-        old = db.engine.execute("SELECT * FROM users WHERE name=%s", req["username"]).first()
-        if len(old) == 0:
-            flash("Cannot edit username")
-            return redirect(url_for("access_manager"))
-        db.engine.execute("UPDATE users SET role_id=%s WHERE name=%s", int(req["role id"]), req["username"])
-    flash("Success")
-    return redirect(url_for("access_manager"))
-
-
-@app.route("/remove_accounts", methods=["POST"])
-@login_required
-@admin_role
-def remove_accounts():
-    complete_request = json.loads(request.get_data().decode("utf-8"))
-    for req in complete_request:
-        db.engine.execute("DELETE FROM users WHERE name=%s", req["username"])
-    return redirect(url_for("access_manager"))
-
-@app.route("/ata_creation", methods=["POST"])
-@login_required
-@admin_role
-def ata_creation():
-    db.engine.execute("INSERT INTO ej(ata_date, president, president_rg, president_cpf, vice_president, vice_president_rg, vice_president_cpf) VALUES(%s, %s, %s, %s, %s, %s, %s)",
-                      request.form["date"], request.form["president"], request.form["president_rg"], request.form["president_cpf"], request.form["vice_president"], request.form["vice_president_rg"], request.form["vice_president_cpf"])
-    return redirect(url_for("ata_manager"))
-
-@app.route("/remove_ata", methods=["POST"])
-@login_required
-@admin_role
-def remove_ata():
-    complete_request = json.loads(request.get_data().decode("utf-8"))
-    for req in complete_request:
-        db.engine.execute("DELETE FROM ej WHERE ata_date=%s", req["ata_date"])
-    return redirect(url_for("ata_manager"))
-
-@app.route("/edit_clients", methods=["POST"])
-@login_required
-@admin_role
-def edit_clients():
-    complete_request = json.loads(request.get_data().decode("utf-8"))
-    for req in complete_request:
-        old = db.engine.execute("SELECT * FROM clients WHERE id=%s", req["id"]).first()
-        if len(old) == 0:
-            return redirect(url_for("access_manager"))
-        db.engine.execute("UPDATE clients SET store_name=%s, address=%s, cep=%s, cnpj=%s, client_name=%s, rg=%s, cpf=%s, email=%s WHERE id=%s", req["store_name"], req["address"], req["cep"], req["cnpj"], req["client_name"], req["rg"], req["cpf"], req["email"], req["id"])
-    return redirect(url_for("access_manager"))
-
-@app.route("/remove_clients", methods=["POST"])
-@login_required
-@admin_role
-def remove_clients():
-    complete_request = json.loads(request.get_data().decode("utf-8"))
-    for req in complete_request:
-        db.engine.execute("UPDATE clients SET removed=true WHERE id=%s", req["id"])
-    return redirect(url_for("access_manager"))
-
-@app.route("/edit_contracts", methods=["POST"])
-@login_required
-@admin_role
-def edit_contracts():
-    complete_request = json.loads(request.get_data().decode("utf-8"))
-    for req in complete_request:
-        old = db.engine.execute("SELECT * FROM services WHERE id=%s", req["id"]).first()
-        if len(old) == 0:
-            return redirect(url_for("access_manager"))
-        descriptions = json.dumps({"short_description": req["short description"], "service_list": req["service list"]})
-        db.engine.execute("UPDATE services SET type=%s, days_to_finish=%s, total_price=%s, payment_price=%s, payment=%s, description=%s client_id=%s WHERE id=%s", req["type"], req["deadline"], req["price"], req["payment price"], req["payments"], descriptions, req["long description"], req["client id"], req["id"])
-    return redirect(url_for("access_manager"))
-
-@app.route("/accept_contracts", methods=["POST"])
-@login_required
-@admin_role
-def accept_contracts():
-    complete_request = json.loads(request.get_data().decode("utf-8"))
-    for req in complete_request:
-        db.engine.execute("UPDATE services SET acceptance_date=now WHERE id=%s", req["id"])
-    return redirect(url_for("access_manager"))
-
-@app.route("/add_payments_contracts", methods=["POST"])
-@login_required
-@admin_role
-def add_payments_contracts():
-    complete_request = json.loads(request.get_data().decode("utf-8"))
-    for req in complete_request:
-        db.engine.execute("INSERT INTO services(payment, price, service_id) VALUES(%s, %s, %s)", req["payment"], req["price"], req["service_id"])
-    return redirect(url_for("access_manager"))
-
-@app.route("/remove_contracts", methods=["POST"])
-@login_required
-@admin_role
-def remove_contracts():
-    complete_request = json.loads(request.get_data().decode("utf-8"))
-    for req in complete_request:
-        db.engine.execute("UPDATE services SET removed=true WHERE id=%s", req["id"])
-    return redirect(url_for("access_manager"))
