--- cerebro-temp/templates/ata_manager.html
@@ -1,76 +1,0 @@
- {% extends "layout.html" %}
-{% block title%}
-Ata manager
-{% endblock %}
-
-{% block main %}
-     <div id="atas-overview">
-         <form action="/ata_creation" method="post">
-
-            <div class="form-row">
-                <div class="form-group">
-                    <label for="date">Date</label>
-                    <input type="date" id="date" name="date" class="form-control">
-                </div>
-
-                <div class="form-group">
-                    <label for="president">President</label>
-                    <input type="text" id="president" name="president" class="form-control">
-                </div>
-
-                <div class="form-group">
-                    <label for="president_rg">President rg</label>
-                    <input type="text" id="president_rg" name="president_rg" class="form-control">
-                </div>
-
-                <div class="form-group">
-                    <label for="president_cpf">President cpf</label>
-                    <input type="text" id="president_cpf" name="president_cpf" class="form-control">
-                </div>
-
-                <div class="form-group">
-                    <label for="vice_president">Vice president</label>
-                    <input type="text" id="vice_president" name="vice_president" class="form-control">
-                </div>
-
-                <div class="form-group">
-                    <label for="vice_president_rg">Vice president rg</label>
-                    <input type="text" id="vice_president_rg" name="vice_president_rg" class="form-control">
-                </div>
-
-                <div class="form-group">
-                    <label for="vice_president_cpf">Vice president cpf</label>
-                    <input type="text" id="vice_president_cpf" name="vice_president_cpf" class="form-control">
-                </div>
-
-            </div>
-            <input type="submit" value="Create ata" id="create_ata">
-        </form>
-            <table class="table table-hover" id="ata">
-                <thead>
-                    <tr>
-                        <th>ata_date</th>
-                        <th>president</th>
-                        <th>president_rg</th>
-                        <th>president_cpf</th>
-                        <th>vice_president</th>
-                        <th>vice_president_rg</th>
-                        <th>vice_president_cpf</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {% for ata in atas %}
-                        <tr class="clickable">
-                            {% for data in ata %}
-                                <td>{{ data }}</td>
-                            {% endfor %}
-                            <td class="edition">
-                                <button class="btn btn-danger btn-sm remove">Remove</button>
-                            </td>
-                        </tr>
-                    {% endfor %}
-                </tbody>
-            </table>
-            <input type="submit" class="save" value="Save">
-        </div>
- {% endblock %}
