--- cerebro-temp/templates/contracts_manager.html
@@ -1,45 +1,0 @@
- {% extends "layout.html" %}
-{% block title%}
-Contracts manager
-{% endblock %}
-
-{% block main %}
-    <div id="contracts-overview">
-        <table class="table table-hover" id="contracts">
-            <thead>
-                <tr>
-                    <th>id</th>
-                    <th>user</th>
-                    <th class="editable" type="text">type</th>
-                    <th class="editable" type="number">deadline</th>
-                    <th class="editable" type="float">price</th>
-                    <th class="editable" type="float">payment price</th>
-                    <th class="editable" type="number">payments</th>
-                    <th class="editable" type="text">short description</th>
-                    <th class="editable" type="text">service list</th>
-                    <th>generation date</th>
-                    <th>acceptance date</th>
-                    <th>first payment date</th>
-                    <th class="editable" type="number">client id</th>
-                    <th>current payments</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for contract in contracts %}
-                    <tr class="clickable">
-                        {% for data in contract %}
-                            <td>{{ data }}</td>
-                        {% endfor %}
-                        <td class="edition">
-                            <button class="btn btn-light btn-sm edit">Edit</button>
-                            <button class="btn btn-info btn-sm accept">Accept</button>
-                            <button class="btn btn-success btn-sm add_payment">Add payment</button>
-                            <button class="btn btn-danger btn-sm remove">Remove</button>
-                        </td>
-                    </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-        <input type="submit" class="save" value="Save">
-    </div>
- {% endblock %}
