--- cerebro-temp/templates/index.html
@@ -1,13 +1,0 @@
-{% extends "layout.html" %}
-{% block title%}
-Gerador contrato
-{% endblock %}
-
-{% block main %}
-    <div>
-        <p>Number of contracts: {% if contracts %}{{ contracts }}{% else %}0{% endif %}</p>
-        <p>Number of contracts with no payments: {% if no_payments_contracts %}{{ no_payments_contracts }}{% else %}0{% endif %}</p>
-        <p>Number of contracts not accepted: {% if no_accepted_contracts %}{{ no_accepted_contracts }}{% else %}0{% endif %}</p>
-        <p>Number of clients: {% if clients %}{{ clients }}{% else %}0{% endif %}</p>
-    </div>
-{% endblock %}
