+++ templates/access.html
@@ -1,14 +1,18 @@
 {% extends "layout.html" %}
 {% block title%}
-Gerador contrato
+Access
 {% endblock %}
 
 {% block main %}
-<form action="/access" method="POST">
-    <div class="form-group">
-        <label for="key">Chave</label>
-        <input type="password" name="key" class="form-control" required>
-    </div>
-    <input type="submit" value="Acessar">
-</form>
+    <form action="/access" method="POST">
+        <div class="form-group">
+            <label for="name">Username</label>
+            <input type="text" id="name" name="name" class="form-control" required>
+        </div>
+        <div class="form-group">
+            <label for="key">Password</label>
+            <input type="password" id="key" name="key" class="form-control" required>
+        </div>
+        <input type="submit" value="Acessar">
+    </form>
 {% endblock %}
